flask_jwt_app/requirements.txt
flask 

PyJWT